split utils into plots and math
