+++ tests/unit_tests/analysis/test_u_model.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from ipsuite.analysis.model.utils import force_decomposition
+from ipsuite.analysis.model.math import force_decomposition
 from ipsuite.geometry import BarycenterMapping
